added my hotels view and added e2e test
